+++ src/Container/ContainerInterface.php
@@ -7,16 +7,10 @@ namespace Rougin\Slytherin\Container;
  *
  * An interface for handling third party dependency injection containers.
  *
- * PSR-11 includes how to get instances from containers using the "get" method.
- * But they don't have yet a unified implementation for defining objects. With
- * this problem, Slytherin will provide an interface that includes a "set"
- * method and this can be served as a "temporary" workaround until the said
- * package, "container-interop/container-interop", will be updated for this matter.
- *
  * @package Slytherin
  * @author  Rougin Royce Gutib <rougingutib@gmail.com>
  */
-interface ContainerInterface extends \Interop\Container\ContainerInterface
+interface ContainerInterface extends \Psr\Container\ContainerInterface
 {
     /**
      * Sets a new instance to the container.