+++ src/Container/Exception/ContainerException.php
@@ -3,13 +3,13 @@
 namespace Rougin\Slytherin\Container\Exception;
 
 /**
- * Not Found Exception
+ * Container Exception
  *
  * A specified exception in handling errors in containers.
  *
  * @package Slytherin
  * @author  Rougin Royce Gutib <rougingutib@gmail.com>
  */
-class NotFoundException extends \InvalidArgumentException implements \Psr\Container\NotFoundExceptionInterface
+class ContainerException extends \Exception implements \Psr\Container\ContainerExceptionInterface
 {
 }