+++ src/Container/VanillaContainer.php
@@ -5,7 +5,7 @@ namespace Rougin\Slytherin\Container;
 /**
  * Container
  *
- * A simple container that is implemented on \Interop\Container\ContainerInterface.
+ * A simple container that is implemented on \Psr\Container\ContainerInterface.
  * NOTE: To be removed in v1.0.0
  *
  * @package Slytherin