+++ tests/IoC/Auryn/ContainerTest.php
@@ -40,7 +40,7 @@ class ContainerTest extends \PHPUnit_Framework_TestCase
             $this->markTestSkipped('Auryn is not installed.');
         }
 
-        if (! interface_exists('Interop\Container\ContainerInterface')) {
+        if (! interface_exists('Psr\Container\ContainerInterface')) {
             $this->markTestSkipped('Container Interop is not installed.');
         }